+++ CHANGELOG.md
@@ -42,6 +42,10 @@ All notable changes to the LoLLMs Platform will be documented in this file.
 
 - feat(notebook): add tab support and extend creation schema
 
+## [2026-03-26 00:59]
+
+- """Generate commit message from summarized changes."""
+
 ## [2026-03-22 22:17]
 
 - chore(deps): bump lollms-client from 1.12.6 to 1.12.9